win_opengl: Split shader loading to it's own file.
