+++ src/win/Makefile.mingw
@@ -451,7 +451,7 @@ ifeq ($(WX), y)
  UIOBJ		:= wx_main.o wx_ui.o wx_stbar.o wx_render.o
 else
  UIOBJ		:= win_ui.o win_stbar.o \
-		   win_sdl.o win_opengl.o glad.o \
+		   win_sdl.o win_opengl.o win_opengl_glslp.o glad.o \
 		   win_dialog.o win_about.o \
 		   win_settings.o win_devconf.o win_snd_gain.o win_specify_dim.o \
 		   win_new_floppy.o win_jsconf.o win_media_menu.o
